requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiohttp-apispec==2.2.3
aiosignal==1.3.1
ajsonrpc==1.2.0
anyio==3.6.1
apispec==3.3.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.2
backcall==0.2.0
bash_kernel==0.8.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
colorama==0.4.4
comm==0.1.2
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
dnspython==2.3.0
ecdsa==0.18.0
eduvpn-client==3.1.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.6.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
gdown==4.7.1
h11==0.13.0
httplib2==0.20.2
idna==2.10
ifaddr==0.2.0
importlib-metadata==6.0.0
inflect==6.0.2
iniconfig==2.0.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libnacl==1.8.0
lockfile==0.12.2
lxml==4.9.1
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
orjson==3.8.6
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
pony==0.7.16
portend==3.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.13.0
pyipv8==2.10.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qrcode==7.4.2
qtconsole==5.3.2
QtPy==2.2.0
rbase==0.1.3
redis==4.4.2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==1.2.1
scikit-learn-extra==0.3.0
scipy==1.10.0
SecretStorage==3.3.1
semantic-version==2.10.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.20.4
tempora==5.2.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.5
uvicorn==0.18.3
wadllib==1.3.6
wcwidth==0.2.5
webargs==5.5.3
webencodings==0.5.1
widgetsnbextension==4.0.3
wsproto==1.1.0
xdg==5.0.0
yappi==1.4.0
yarl==1.8.2
zc.lockfile==2.0
zeroconf==0.39.0
zipp==1.0.0